+++ cmd/serve.go
@@ -17,7 +17,6 @@ package cmd
 import (
 	"context"
 	"net/http"
-	"os"
 
 	"code.cloudfoundry.org/lager"
 	"github.com/GoogleCloudPlatform/gcp-service-broker/brokerapi/brokers"
@@ -28,6 +27,7 @@ import (
 	"github.com/GoogleCloudPlatform/gcp-service-broker/pkg/compatibility"
 	"github.com/GoogleCloudPlatform/gcp-service-broker/pkg/server"
 	"github.com/GoogleCloudPlatform/gcp-service-broker/pkg/toggles"
+	"github.com/GoogleCloudPlatform/gcp-service-broker/utils"
 	"github.com/pivotal-cf/brokerapi"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -62,9 +62,7 @@ func init() {
 
 func serve() {
 
-	logger := lager.NewLogger("gcp-service-broker")
-	logger.RegisterSink(lager.NewWriterSink(os.Stderr, lager.ERROR))
-	logger.RegisterSink(lager.NewWriterSink(os.Stdout, lager.DEBUG))
+	logger := utils.NewLogger("gcp-service-broker")
 
 	models.ProductionizeUserAgent()