+++ cmd/show.go
@@ -18,7 +18,6 @@ import (
 	"fmt"
 	"log"
 
-	"code.cloudfoundry.org/lager"
 	"github.com/GoogleCloudPlatform/gcp-service-broker/brokerapi/brokers/models"
 	"github.com/GoogleCloudPlatform/gcp-service-broker/db_service"
 	"github.com/GoogleCloudPlatform/gcp-service-broker/utils"
@@ -58,7 +57,7 @@ func addDumpTableCommand(parent *cobra.Command, name string, value interface{}) 
 }
 
 func tableToJson(results interface{}) {
-	logger := lager.NewLogger("show-command")
+	logger := utils.NewLogger("show-command")
 	db := db_service.SetupDb(logger)
 
 	if err := db.Find(results).Error; err != nil {